Move interface to a dedicated package.
