+++ api/metrics-api-jaxrs/src/main/java/org/hawkular/metrics/api/jaxrs/handler/template/IMetricsHandler.java
@@ -14,7 +14,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.hawkular.metrics.api.jaxrs.handler;
+package org.hawkular.metrics.api.jaxrs.handler.template;
 
 import java.util.List;
 import java.util.Map;
